js: minimize and reformat imported StableLib code
- Specialize to AES
- Indent at 2 spaces
- Separate src/ test/ and bench/ directories
- Use mocha-typescript and chai for tests
- Import any missing StableLib code needed to make the tests pass
