+++ js/package.json
@@ -1,26 +1,38 @@
 {
-  "name": "@stablelib/ctr",
-  "version": "0.1.2",
-  "description": "Counter block cipher mode (CTR)",
-  "main": "./dist/ctr.js",
-  "typings": "./dist/ctr.d.ts",
-  "author": "Dmitry Chestnykh",
+  "name": "sivchain",
+  "version": "0.0.0",
+  "description": "Symmetric encryption using the AES-SIV (RFC 5297) and CHAIN constructions",
+  "homepage": "https://github.com/zcred/sivchain/tree/master/js/",
+  "main": "index.js",
+  "author": "Tony Arcieri",
   "license": "MIT",
-  "publishConfig": {
-    "access": "public"
+  "repository": {
+    "type": "git",
+    "url": "git+https://github.com/zcred/sivchain.git"
+  },
+  "keywords": [
+    "cryptography",
+    "encryption",
+    "streaming",
+    "security"
+  ],
+  "bugs": {
+    "url": "https://github.com/zcred/sivchain/issues"
+  },
+  "directories": {
+    "test": "test"
   },
   "scripts": {
-    "build": "tsc",
-    "test": "jasmine JASMINE_CONFIG_PATH=../../configs/jasmine.json",
-    "bench": "node ./dist/ctr.bench.js"
-  },
-  "dependencies": {
-    "@stablelib/blockcipher": "^0.1.2",
-    "@stablelib/wipe": "^0.0.1"
+    "test": "mocha --require ts-node/register test/**/*.spec.ts"
   },
   "devDependencies": {
-    "@stablelib/aes": "^0.1.2",
-    "@stablelib/benchmark": "^0.0.1",
-    "@stablelib/hex": "^0.0.1"
+    "chai": "^3.5.0",
+    "mocha": "^3.2.0",
+    "mocha-typescript": "^1.0.23",
+    "ts-node": "^3.0.2",
+    "typescript": "^2.2.1",
+    "@types/chai": "^3.4.35",
+    "@types/mocha": "^2.2.40",
+    "@types/node": "^7.0.13"
   }
 }
