+++ js/test/wipe.spec.ts
@@ -1,0 +1,11 @@
+import { suite, test } from "mocha-typescript";
+import { expect } from "chai";
+import { wipe } from "../src/wipe";
+
+@suite class WipeSpec {
+  @test "should wipe bytes"() {
+    const a = new Uint8Array([1, 2, 3, 4]);
+    wipe(a);
+    expect(a).to.eql(new Uint8Array(4));
+  }
+}
