+++ js/tsconfig.json
@@ -1,10 +1,20 @@
 {
-    "extends": "../../configs/tsconfig",
     "compilerOptions": {
-        "outDir": "dist"
-    },
-    "exclude": [
-        "node_modules",
-        "dist"
-    ]
+        "target": "ES5",
+        "module": "commonjs",
+        "removeComments": false,
+        "preserveConstEnums": true,
+        "sourceMap": true,
+        "experimentalDecorators": true,
+        "declaration": true,
+        "pretty": true,
+        "alwaysStrict": true,
+        "noFallthroughCasesInSwitch": true,
+        "noImplicitAny": true,
+        "noImplicitReturns": true,
+        "noImplicitThis": true,
+        "noUnusedParameters": true,
+        "strictNullChecks": true,
+        "newLine": "LF"
+    }
 }
